learn/06-observability/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0
